+++ iris/service/content/petition/service.py
@@ -1,3 +1,4 @@
+import time
 import transitions
 
 from transitions.extensions.nesting import NestedState
@@ -70,11 +71,12 @@ class PetitionsRESTMapper(rest.DocumentRESTMapperMixin,
     """A mapper for the petitions admin REST API
     """
 
-    NAME = 'petitions'
+    NAME = PETITIONS_MAPPER_NAME
 
     DOC_CLASS = Petition
 
     QUERY_PARAMS = {
+        'created_range': queries.rangeFilter('dc.created'),
         'ft': queries.fulltextQuery(['tags_ft',
                                      'title_ft',
                                      'description_ft',
@@ -105,6 +107,9 @@ class PetitionsRESTMapper(rest.DocumentRESTMapperMixin,
         'score': queries.scoreSorter,
         'default': queries.fieldSorter('dc.created', 'DESC'),
     }
+
+    TRENDING_STEP_START = 7  # days
+    TRENDING_STEP_MAX = 12 * 7
 
     def event(self, contentId, transitionName, data={},
               resolve=[], extend=[]):
@@ -185,6 +190,90 @@ class PetitionsRESTMapper(rest.DocumentRESTMapperMixin,
             # The owners email is trusted so we can also trust the email
             # on the relation.
             doc.owner = {'email_trusted': True}
+
+    DEFAULT_TRENDING = (7,   # look back days
+                        90,  # max look back days
+                        7    # increment
+                       )
+
+    def search(self,
+               offset=0,
+               limit=10,
+               sort=None,
+               testing_only=False,
+               debug=None,
+               trending=None,
+               **params):
+        """Extends the search endpoint with the special `trending` sort.
+
+        It is possible to provide `sort=trending` which will then
+        sort by trending petitions.
+        If this sort is used all parameters except the `limit` are ignored.
+
+        Trending petitions is implemented by searching for petitions with the
+        most supports in the last `look back days` (see DEFAULT_TRENDING).
+        If the query provides less than `limit` petitions the time range is
+        extended by `increment`. This is repeated until `max look back days`
+        is reached.
+
+        For testing it is also possible to provide the trending parameters as
+        `trending` query parameter. The parameter must provide a list with
+        exactly 3 integer values. The values are used the same way as
+        DEFAULT_TRENDING is defined.
+        Example: trending=14,120,14
+        """
+        if (trending is None
+            and (sort is None
+                 or 'trending' not in sort)
+           ):
+            # use the standard search
+            return super(PetitionsRESTMapper, self).search(
+                offset, limit, sort, testing_only, debug, **params)
+        if trending is None:
+            # provide trending because of the sort
+            trending = self.DEFAULT_TRENDING
+        try:
+            current, tr_max, tr_step = trending
+        except:
+            raise ValueError(
+                'Wrong number (expected 3)'
+                ' of parameters in trending: "%s"' % trending)
+        while True:
+            fromTs = (int(time.time()) - 86400 * current) * 1000
+            query = queries.rangeFilter('dc.created', ['gte'])([fromTs])
+            aggregations = {
+                "trending": {
+                    "terms": {
+                        "field": "relations.petition",
+                        "size": limit
+                    }
+                }
+            }
+            body = {
+                "from": 0,
+                "size": 0,
+                "query": query,
+                "aggregations": aggregations,
+            }
+            searchresult = Supporter.search(body)
+            buckets = searchresult['aggregations']['trending']['buckets']
+            if tr_step > 0 and current < tr_max and len(buckets) < limit:
+                current += tr_step
+                if current > tr_max:
+                    # don't go above the requested days
+                    current = tr_max
+                # next time use a bigger step
+                tr_step *= 2
+                continue
+            data = []
+            keys = [b['key'] for b in buckets[:limit]]
+            if keys:
+                data = self.get(keys)
+            result = {
+                'data': data,
+                'total': len(data),
+            }
+            return result
 
 
 @RestService("petition_public_api")